try fix github actions the action to install hub CLI is no longer automatically compatible with the current github actions.
The current github actions discourages the use of  and the command
is only allow if we turn on an unsafe flag.

consequently, on the action's github issues page, there exist a claim that
hub cli commands are available in actions without installing. Thus this commit removes
the installation of hub cli to test this claim